LayoutTests/inspector/debugger/breakpoint-scope-expected.txt

CONSOLE MESSAGE: line 1: Paused at line: 4, column: 8
Testing that we can access scope in various functions.

Starting Test
Hit breakpoint at line: 4, column: 8
scope-chain-type-local properties:
    resolve
    reject
scope-chain-type-closure properties:
    p
scope-chain-type-globalLexicalEnvironment properties:
    lexicalVariable
scope-chain-type-global (properties not listed)
Tests done
